web assets go project structure
