--- src/site/go.mod
@@ -1,3 +1,0 @@
-module site
-
-go 1.21.1
