+++ web/scripts/util.js
